Added comments in kode.go

+++ kode/kode.go
@@ -10,8 +10,8 @@ func Run(code string) error {
 	}
 
 	// Create a new main scope.
-	_debug := CreateVariable(false)
-	_max_recursion := CreateVariable(5000)
+	_debug := CreateVariable(false)               // DEBUG variable prints debug info to console
+	_max_recursion := CreateVariable(int64(5000)) // Max recursion depth for functions
 	scope := CreateFunction([]Argument{}, map[string]*Variable{"_DEBUG": &_debug, "_MAX_RECURSION": &_max_recursion}, "null", nil, strings.ReplaceAll(code, "\r", " "))
 
 	// Enter the main scope.
